finish rewriting DiaryHelper
functions kept as-is; except some syntax modernized

With all of its functions now in diaryHelper.ts, the DiaryHelper factory can be removed altogether
